unify abi encoding benches
