+++ benches/benches/u256.rs
@@ -26,7 +26,7 @@ fn u256_benchmark(c: &mut Criterion) {
 
     let mut group1 = c.benchmark_group("U256 Operations");
 
-    group1.bench_function("Ethers/get-amount-in", |b| {
+    group1.bench_function("Ethers/getAmountIn", |b| {
         b.iter(|| {
             _ = e_get_amount_in(
                 black_box(e_uniswap_pair.reserve0),
@@ -38,7 +38,7 @@ fn u256_benchmark(c: &mut Criterion) {
         })
     });
 
-    group1.bench_function("Alloy/get-amount-in", |b| {
+    group1.bench_function("Alloy/getAmountIn", |b| {
         b.iter(|| {
             _ = a_get_amount_in(
                 black_box(a_uniswap_pair.reserve0),
@@ -50,7 +50,7 @@ fn u256_benchmark(c: &mut Criterion) {
         })
     });
 
-    group1.bench_function("Ethers/get-amount-out", |b| {
+    group1.bench_function("Ethers/getAmountOut", |b| {
         b.iter(|| {
             _ = e_get_amount_out(
                 black_box(e_uniswap_pair.reserve0),
@@ -60,7 +60,7 @@ fn u256_benchmark(c: &mut Criterion) {
         })
     });
 
-    group1.bench_function("Alloy/get-amount-out", |b| {
+    group1.bench_function("Alloy/getAmountOut", |b| {
         b.iter(|| {
             _ = a_get_amount_out(
                 black_box(a_uniswap_pair.reserve0),
